--- src/main/java/com/udc/converter/ChartTypeConverter.java
@@ -1,12 +1,0 @@
-package com.udc.converter;
-
-import com.udc.model.Constant;
-
-import javax.faces.convert.FacesConverter;
-
-/**
- * Created by sdrahnea
- */
-@FacesConverter("chartTypeConverter")
-public class ChartTypeConverter extends AbstractConverter<Constant> {
-}
