--- src/main/java/com/udc/converter/ui/ConstantConverter.java
@@ -1,12 +1,0 @@
-package com.udc.converter.ui;
-
-import com.udc.model.Constant;
-
-import javax.faces.convert.FacesConverter;
-
-/**
- * Created by sdrahnea
- */
-@FacesConverter("constantConverter")
-public class ConstantConverter extends AbstractConverter<Constant> {
-}
