--- src/main/java/com/udc/converter/ui/DashboardConverter.java
@@ -1,12 +1,0 @@
-package com.udc.converter.ui;
-
-import com.udc.model.Constant;
-
-import javax.faces.convert.FacesConverter;
-
-/**
- * Created by sdrahnea
- */
-@FacesConverter("dashboardConverter")
-public class DashboardConverter extends AbstractConverter<Constant> {
-}
