+++ src/main/java/com/udc/converter/ui/UIDashboardConverter.java
@@ -1,4 +1,4 @@
-package com.udc.converter;
+package com.udc.converter.ui;
 
 import com.udc.model.Constant;
 
@@ -8,5 +8,5 @@ import javax.faces.convert.FacesConverter;
  * Created by sdrahnea
  */
 @FacesConverter("dashboardConverter")
-public class DashboardConverter extends AbstractConverter<Constant> {
+public class UIDashboardConverter extends UIAbstractConverter<Constant> {
 }
